Task 0 - Pedido Livraria

+++ 0x06/pedido_livraria/Pedido.java
@@ -4,7 +4,7 @@ public class Pedido {
     private double percentualDesconto;
     private ItemPedido[] itens;
 
-    public Pedido(double percentualDesconto, ItemPedido[] itens) {
+    public Pedido(int percentualDesconto, ItemPedido[] itens) {
         this.percentualDesconto = percentualDesconto;
         this.itens = itens;
     }
